% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{unite}
\alias{unite}
\title{A \code{unite} Function}
\usage{
unite(col1, col2)
}
\arguments{
\item{col1}{column 1}

\item{col2}{column 2}

\item{data}{data you are working with}
}
\description{
This function allows you to combine two columns into where the variables are the same just with different names
And you want to combine them because one column has NA in rows where the other column is filled
I came across these things a lot working with Monitoring the Future data which is absolute trash data
to work with
}
\examples{
unite()


Lets say the variables V1112 and V1113 both are variables for "how likely is it you will graduate college?"
They were asked on different forms hence different variable names. We want to combine them into one
}
